Update redux module naming
