--- src/redux-modules/data-globe/data-globe.js
@@ -1,8 +1,0 @@
-import reducerRegistry from 'reducerRegistry';
-
-import * as actions from './data-globe-actions';
-import reducers, { initialState } from './data-globe-reducers';
-
-const reduxConfig = { actions, reducers, initialState };
-
-export default reducerRegistry.registerModule('dataGlobeSpec', reduxConfig);
