+++ src/redux-modules/featured-maps-list/featured-maps-list.js
@@ -1,0 +1,10 @@
+import reducerRegistry from 'reducerRegistry';
+
+import * as actions from './featured-maps-list-actions';
+import reducers, { initialState } from './featured-maps-list-reducers';
+
+export const reduxConfig = { actions, reducers, initialState };
+
+reducerRegistry.registerModule('featuredMapsList', reduxConfig);
+
+export default actions;
